+++ tests/origio/pipeline/test_example.py
@@ -33,7 +33,6 @@ def test_resource_example(
         assert errors.message == "12314 is not of type 'string'"
 
     instance: PipelineInstance = instances[0]
-    instance.schemaId = "blabla"
     instance.create()
 
     assert mock_create_pipeline_instance.called
@@ -59,10 +58,8 @@ def test_bootstrap_script(
         sdk,
         id="uuid?",
         datasetUri="boligpriser",
-        schemaId="coming soon",
         taskConfig="transformation goes here",
         pipelineArn=pipeline.arn,
-        useLatestEdition=True,
     )
     valid, error = instance.validate()
     assert not valid