+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * `PipelineInstance` now takes an optional parameter `pipelineProcessorId`,
   intended to supersede `pipelineArn` once all users have been updated. The
   `pipelineArn` parameter is now optional.
+* Support the new status API response format. This affects the upload command response, now returning a `trace_id` key instead of `status`.
 
 ## 0.2.7